requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asyncstdlib==3.10.5
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
discord==2.0.0
discord-pretty-help==1.3.4
discord-py-slash-command==4.2.1
discord.py==2.0.1
frozenlist==1.3.1
html5lib==1.1
idna==3.4
multidict==6.0.2
pycparser==2.21
PyNaCl==1.5.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
yarl==1.8.1
youtube-dl==2021.12.17
youtube-search==2.1.2
